Aide: First attempt at creating a settings page and letting user choose date format
# Aider chat conversation:

# Context:
USER: This is a Chrome extension. Do you understand how it saves and fecthes stored information?
USER: Good. Currently, it uses the date format Tuesday, 1 August 2023 and Wed, 6 Dec 2023. This is what I like personally but there should be a settings page where the user can choose the format of the date. Maybe they can choose from a list of common date formats