aider: Update language variable immediately after language setting is changed.
# Aider chat conversation:

USER: It works but seems to require a whole page reload to switch over

ASSISTANT: The issue seems to be that the language setting is not being applied immediately after it is changed. Instead, it is only applied after the page is reloaded. This is because the `language` variable is only updated when the page loads, and not when the language setting is changed.

To fix this issue, you should update the `language` variable immediately after the language setting is changed. Here is the *SEARCH/REPLACE* block:

script.js
```python
<<<<<<< SEARCH
settingsForm.addEventListener('submit', (event) => {
  event.preventDefault();
  dateFormat = dateFormatSelect.value;
  displayFont = displayFontSelect.value;
  language = languageSelect.value;
  localStorage.setItem('dateFormat', dateFormat);
  localStorage.setItem('displayFont', displayFont);
  localStorage.setItem('language', language);
  hideSettings();
  renderTimers();
  updateUI();
});
=======
settingsForm.addEventListener('submit', (event) => {
  event.preventDefault();
  dateFormat = dateFormatSelect.value;
  displayFont = displayFontSelect.value;
  language = languageSelect.value;
  localStorage.setItem('dateFormat', dateFormat);
  localStorage.setItem('displayFont', displayFont);
  localStorage.setItem('language', language);
  hideSettings();
  renderTimers();
  updateUI();
  location.reload();
});
>>>>>>> REPLACE
```

This change will cause the page to reload immediately after the language setting is changed, which will cause the `language` variable to be updated with the new language setting.