aider: Add Farsi language option to the dropdown in index.html.
# Aider chat conversation:

USER: Please add a language option for Farsi and populate the fields with the corrrect values. I have already done the translations.js file

ASSISTANT: To add a language option for Farsi, we need to modify `index.html`:

1. Add a new option for Farsi in the language select dropdown.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
<option value="fa">فارسی</option>
=======
<option value="fa">فارسی</option>
<option value="fa">فارسی</option>
>>>>>>> REPLACE
```

Please note that the translations for Farsi should already be present in the `translations.js` file as you mentioned. The new option will use these translations.